Rename to getHubLabBucketItems
