+++ src/extension.ts
@@ -28,7 +28,7 @@ import checkIfUserStarred from "./utils/github/checkIfUserStarred";
 import getMostRelevantJiraTickets from "./utils/jira/getMostRelevantJiraTickets";
 import { WatermelonAuthenticationProvider } from "./auth";
 import { ContextItem } from "./ContextItem";
-import { getHubLabBucketItems } from "./utils/treeview/getGitHubItems";
+import { getHubLabBucketItems } from "./utils/treeview/getHubLabBucketItems";
 import { getGitItems } from "./utils/treeview/getGitItems";
 import { getJiraItems } from "./utils/treeview/getJiraItems";
 import { getSlackItems } from "./utils/treeview/getSlackItems";
